+++ Source/Krypton Toolkit/Krypton.Toolkit.Suite.Extended.Messagebox/Controls Visuals/KryptonMessageBoxExtendedFormTest.cs
@@ -1,0 +1,10 @@
+﻿namespace Krypton.Toolkit.Suite.Extended.Messagebox
+{
+    public partial class KryptonMessageBoxExtendedFormTest : KryptonForm
+    {
+        public KryptonMessageBoxExtendedFormTest()
+        {
+            InitializeComponent();
+        }
+    }
+}