+++ Source/Krypton Toolkit/Krypton.Toolkit.Suite.Extended.Navigator/Classes/OutlookBarButton.cs
@@ -6,10 +6,6 @@
  */
 #endregion
 
-using System;
-using System.ComponentModel;
-using System.Drawing;
-
 namespace Krypton.Toolkit.Suite.Extended.Navigator
 {
     [DesignTimeVisible(false), DefaultProperty("Text")]