+++ Source/Krypton Toolkit/Krypton.Toolkit.Suite.Extended.Navigator/Controls/KryptonTabControlLite.cs
@@ -6,11 +6,6 @@
  */
 #endregion
 
-using System;
-using System.Drawing;
-using System.Drawing.Drawing2D;
-using System.Windows.Forms;
-
 namespace Krypton.Toolkit.Suite.Extended.Navigator
 {
     [ToolboxBitmap(typeof(TabControl))]