+++ Source/Krypton Toolkit/Krypton.Toolkit.Suite.Extended.Networking/Classes/General/NetworkNodeBrowser.cs
@@ -6,8 +6,6 @@
  */
 #endregion
 
-using Krypton.Toolkit.Suite.Extended.Developer.Utilities;
-
 namespace Krypton.Toolkit.Suite.Extended.Networking
 {
     /// <summary>